+++ packages/runtime-tools/src/coding/tools/followup-task/index.ts
@@ -1,0 +1,13 @@
+export { FOLLOWUP_TASK_TOOL_DESCRIPTION } from "./description.js";
+export {
+	createFollowupTaskTool,
+	type FollowupTaskToolInput,
+	FollowupTaskToolInputSchema,
+	type FollowupTaskToolOptions,
+} from "./followup-task-tool.js";
+export {
+	createFollowupTaskToolRegistration,
+	FOLLOWUP_TASK_TOOL_CATEGORY,
+	FOLLOWUP_TASK_TOOL_SCOPES,
+	type FollowupTaskToolRegistrationOptions,
+} from "./registration.js";